+++ app/__init__.py
@@ -1,0 +1,5 @@
+"""
+Ontario Damages Compendium Application Package
+"""
+
+__version__ = "1.0.0"
