+++ app/core/__init__.py
@@ -1,0 +1,7 @@
+"""
+Core business logic for the Ontario Damages Compendium
+"""
+
+from .config import *
+from .search import *
+from .data_loader import *
