+++ app/ui/__init__.py
@@ -1,0 +1,5 @@
+"""
+UI components and visualizations
+"""
+
+from .visualizations import *
