+++ streamlit_app.py
@@ -1,25 +1,28 @@
 """
 Ontario Damages Compendium - Visual Search Tool
 A professional legal tool for searching comparable personal injury awards
+
+Main application file - handles UI and user interactions only.
+Business logic has been moved to app/ modules for better organization.
 """
 
 import streamlit as st
-import json
 import numpy as np
-from pathlib import Path
-from sentence_transformers import SentenceTransformer
-from sklearn.metrics.pairwise import cosine_similarity
-import re
 import tempfile
 import os
 from datetime import datetime
-from typing import Optional, List, Dict, Tuple, Any
+from typing import Dict, List
 
-# Import custom modules
+# Import custom modules (refactored into app/ package)
+from app.core.config import *
+from app.core.data_loader import initialize_data
+from app.core.search import search_cases, extract_damages_value
+from app.ui.visualizations import create_inflation_chart, calculate_chart_statistics
+
+# Import other application modules
 from expert_report_analyzer import analyze_expert_report
 from pdf_report_generator import generate_damages_report
 from inflation_adjuster import (
-    adjust_for_inflation,
     DEFAULT_REFERENCE_YEAR,
     get_data_source,
     get_cpi_data,
@@ -27,33 +30,10 @@ from inflation_adjuster import (
     reload_cpi_data
 )
 
-# Plotly for interactive charts
-import plotly.graph_objects as go
+# =============================================================================
+# PAGE CONFIGURATION
+# =============================================================================
 
-# ============================================================================
-# CONFIGURATION CONSTANTS
-# ============================================================================
-
-# Search algorithm weights
-EMBEDDING_WEIGHT = 0.7  # Weight for semantic similarity (70%)
-REGION_WEIGHT = 0.3     # Weight for anatomical region matching (30%)
-
-# Damage value filtering
-MIN_DAMAGE_VALUE = 1000        # Minimum reasonable damage award
-MAX_DAMAGE_VALUE = 10_000_000  # Maximum reasonable damage award
-
-# Display settings
-DEFAULT_TOP_N_RESULTS = 15      # Default number of search results
-CHART_MAX_CASES = 15            # Maximum cases to show in charts
-CASE_SUMMARY_MAX_LENGTH = 400   # Max characters for case summary display
-EXPANDED_RESULTS_COUNT = 3      # Number of results expanded by default
-
-# Model settings
-EMBEDDING_MODEL_NAME = "sentence-transformers/all-MiniLM-L6-v2"
-
-# ============================================================================
-
-# Page configuration
 st.set_page_config(
     page_title="Ontario Damages Compendium",
     page_icon="⚖️",
@@ -108,12 +88,6 @@ st.markdown("""
         color: #6366f1;
         font-weight: 600;
     }
-    .body-map-container {
-        display: flex;
-        justify-content: center;
-        gap: 2rem;
-        margin: 2rem 0;
-    }
     .instructions {
         background-color: #eff6ff;
         border-left: 4px solid #3b82f6;
@@ -124,270 +98,26 @@ st.markdown("""
 </style>
 """, unsafe_allow_html=True)
 
-# Load resources
-@st.cache_resource
-def load_embedding_model() -> SentenceTransformer:
-    """Load the sentence transformer model"""
-    return SentenceTransformer(EMBEDDING_MODEL_NAME)
+# =============================================================================
+# INITIALIZE DATA
+# =============================================================================
 
-@st.cache_data
-def load_cases() -> Optional[List[Dict[str, Any]]]:
-    """Load case data with embeddings"""
-    data_path = Path("data/damages_with_embeddings.json")
-    if not data_path.exists():
-        st.error(f"❌ Data file not found: {data_path}")
-        st.info("Please run the `01_extract_and_embed.ipynb` notebook first to generate the data.")
-        return None
+model, cases, region_map = initialize_data()
 
-    with open(data_path) as f:
-        return json.load(f)
+# =============================================================================
+# INITIALIZE SESSION STATE
+# =============================================================================
 
-@st.cache_data
-def load_region_map() -> Dict[str, Any]:
-    """Load region mapping"""
-    with open("region_map.json") as f:
-        return json.load(f)
+if 'search_results' not in st.session_state:
+    st.session_state.search_results = None
+if 'analysis_data' not in st.session_state:
+    st.session_state.analysis_data = None
 
-# Initialize
-model = load_embedding_model()
-cases = load_cases()
-region_map = load_region_map()
+# =============================================================================
+# MAIN UI
+# =============================================================================
 
-if cases is None:
-    st.stop()
-
-def normalize_region_name(region_name: str) -> Optional[str]:
-    """
-    Normalize region names from case data to standardized region IDs.
-
-    Handles variations in terminology by mapping free-text region names
-    (e.g., "neck", "cervical", "C-spine") to standardized region IDs
-    (e.g., "cervical_spine") using the region_map compendium terms.
-
-    Algorithm:
-    1. Check if region_map key matches (with underscores replaced by spaces)
-    2. Check if any compendium term for a region matches the input
-    3. Return first match found, or None if no match
-
-    Args:
-        region_name: Free-text region name from case data
-
-    Returns:
-        Standardized region ID (str) or None if no match found
-
-    Example:
-        normalize_region_name("neck injury") → "cervical_spine"
-        normalize_region_name("C5-C6") → "cervical_spine"
-    """
-    region_lower = region_name.lower()
-
-    # Iterate through all known regions in our mapping
-    for key, data in region_map.items():
-        # Check if the region key itself appears (e.g., "cervical_spine" → "cervical spine")
-        if key.replace("_", " ") in region_lower:
-            return key
-
-        # Check all synonyms/clinical terms for this region
-        for term in data["compendium_terms"]:
-            if term.lower() in region_lower:
-                return key
-
-    return None  # No match found
-
-def calculate_region_overlap(case_region: str, selected_regions: List[str]) -> float:
-    """
-    Calculate overlap score between a case's region and user-selected regions.
-
-    Implements a tiered matching system:
-    - Score 1.0 = Perfect match (normalized regions match exactly)
-    - Score 0.7 = Partial match (case region text contains a synonym)
-    - Score 0.0 = No match
-
-    This scoring allows cases with closely related but not identical regions
-    to still appear in results (e.g., "shoulder" matching "rotator cuff").
-
-    Algorithm:
-    1. If no regions selected, return 0.0 (no filtering)
-    2. Normalize the case region using our standardized mapping
-    3. Check for exact match against selected regions → 1.0
-    4. Check if any synonym appears in case region text → 0.7
-    5. Otherwise return 0.0 (no match)
-
-    Args:
-        case_region: Free-text region from case data (e.g., "NECK")
-        selected_regions: List of standardized region IDs user selected
-
-    Returns:
-        Float score between 0.0 and 1.0 indicating match quality
-
-    Example:
-        calculate_region_overlap("Cervical spine injury", ["cervical_spine"]) → 1.0
-        calculate_region_overlap("Neck and shoulder", ["cervical_spine"]) → 0.7
-        calculate_region_overlap("Lower back", ["cervical_spine"]) → 0.0
-    """
-    if not selected_regions:
-        return 0.0  # No regions selected = no regional filtering
-
-    # Normalize the case's region to a standardized ID
-    case_region_normalized = normalize_region_name(case_region)
-
-    # Perfect match: case region maps to one of the selected regions
-    if case_region_normalized in selected_regions:
-        return 1.0
-
-    # Partial match: check if any synonym of selected regions appears in case text
-    case_region_lower = case_region.lower()
-    for region_id in selected_regions:
-        region_data = region_map.get(region_id, {})
-        terms = region_data.get("compendium_terms", [])
-
-        for term in terms:
-            if term.lower() in case_region_lower:
-                return 0.7  # Partial match (synonym found)
-
-    return 0.0  # No match
-
-def search_cases(
-    injury_text: str,
-    selected_regions: List[str],
-    gender: Optional[str] = None,
-    age: Optional[int] = None,
-    top_n: int = DEFAULT_TOP_N_RESULTS
-) -> List[Tuple[Dict[str, Any], float, float]]:
-    """
-    Hybrid search algorithm combining semantic similarity with anatomical region matching.
-
-    This implements a two-stage ranking system:
-
-    STAGE 1 - Region-Based Filtering:
-    - Calculate region overlap scores for all cases
-    - Filter to cases with region_score > 0 (any match)
-    - Fallback to all cases if no region matches found
-
-    STAGE 2 - Hybrid Scoring:
-    - Compute semantic similarity using sentence-transformer embeddings (70% weight)
-    - Combine with anatomical region overlap score (30% weight)
-    - Sort by combined score and return top N results
-
-    ALGORITHM RATIONALE:
-    - Embedding similarity captures semantic meaning and injury descriptions
-    - Region overlap ensures anatomical relevance
-    - 70/30 weighting balances precision (regions) with recall (semantics)
-    - Weight values optimized through validation on historical cases
-
-    QUERY CONSTRUCTION:
-    - Concatenates region labels + injury text for richer context
-    - Example: "Cervical Spine (C1-C7) chronic disc herniation with radiculopathy"
-    - This helps the embedding model understand anatomical context
-
-    Args:
-        injury_text: User's description of the injury (clinical detail encouraged)
-        selected_regions: List of standardized region IDs from sidebar selection
-        gender: Male/Female/None - currently not used in filtering (future enhancement)
-        age: Age of plaintiff - currently not used in filtering (future enhancement)
-        top_n: Number of results to return (default from config)
-
-    Returns:
-        List of (case_dict, embedding_similarity, combined_score) tuples,
-        sorted by combined_score descending
-
-    Example:
-        results = search_cases(
-            "C5-C6 disc herniation with chronic radiculopathy",
-            ["cervical_spine"],
-            top_n=15
-        )
-        # Returns top 15 cases ranked by similarity + region match
-    """
-    # Build query text by prepending region labels to provide anatomical context
-    # This improves embedding quality for medical/legal terminology
-    region_labels = ' '.join([region_map[r]['label'] for r in selected_regions])
-    query_text = f"{region_labels} {injury_text}".strip()
-
-    # Generate embedding vector for the query using cached sentence-transformer model
-    query_vec = model.encode(query_text).reshape(1, -1)
-
-    # ========================================================================
-    # STAGE 1: Region-Based Filtering
-    # ========================================================================
-    # Calculate region overlap scores for each case and filter to matches
-    if selected_regions:
-        filtered_cases = []
-        for case in cases:
-            # Get the region overlap score (0.0, 0.7, or 1.0)
-            region_score = calculate_region_overlap(case.get("region", ""), selected_regions)
-
-            # Only include cases with some regional overlap
-            if region_score > 0:
-                case_copy = case.copy()
-                case_copy["region_score"] = region_score  # Add score to case for later use
-                filtered_cases.append(case_copy)
-
-        # FALLBACK: If no cases match the selected regions at all,
-        # include all cases but with region_score=0 (pure semantic search)
-        # This prevents returning empty results for unusual region combinations
-        if not filtered_cases:
-            filtered_cases = [{**c, "region_score": 0} for c in cases]
-    else:
-        # No regions selected: use all cases with region_score=0
-        # This makes regional weighting irrelevant (pure semantic search)
-        filtered_cases = [{**c, "region_score": 0} for c in cases]
-
-    # ========================================================================
-    # STAGE 2: Semantic Similarity Calculation
-    # ========================================================================
-    # Extract pre-computed embedding vectors for all filtered cases
-    vectors = np.array([c["embedding"] for c in filtered_cases])
-
-    # Calculate cosine similarity between query and each case
-    # Returns array of similarity scores in range [0, 1]
-    embedding_sims = cosine_similarity(query_vec, vectors)[0]
-
-    # ========================================================================
-    # STAGE 3: Hybrid Score Combination
-    # ========================================================================
-    # Combine semantic similarity (70%) with regional matching (30%)
-    # Formula: score = 0.7 * embedding_sim + 0.3 * region_score
-    #
-    # This weighting was chosen because:
-    # - Embeddings capture nuanced injury descriptions (primary signal)
-    # - Region matching ensures anatomical relevance (secondary filter)
-    # - 70/30 split validated through testing on historical cases
-    combined_scores = (EMBEDDING_WEIGHT * embedding_sims +
-                      REGION_WEIGHT * np.array([c["region_score"] for c in filtered_cases]))
-
-    # ========================================================================
-    # STAGE 4: Ranking and Return
-    # ========================================================================
-    # Sort by combined score (highest first) and return top N
-    ranked = sorted(
-        zip(filtered_cases, embedding_sims, combined_scores),
-        key=lambda x: x[2],  # Sort by combined_score (index 2)
-        reverse=True
-    )
-
-    return ranked[:top_n]
-
-def extract_damages_value(case: Dict[str, Any]) -> Optional[float]:
-    """Extract numeric damages value from case"""
-    if case.get("damages"):
-        return case["damages"]
-
-    # Try to extract from raw fields or summary
-    summary = case.get("summary_text", "")
-    dollar_amounts = re.findall(r'\$[\d,]+', summary)
-
-    for amount in dollar_amounts:
-        try:
-            value = float(amount.replace("$", "").replace(",", ""))
-            if MIN_DAMAGE_VALUE <= value <= MAX_DAMAGE_VALUE:
-                return value
-        except:
-            continue
-
-    return None
-
-# Main UI
+# Header
 st.markdown('<div class="main-header">⚖️ Ontario Damages Compendium</div>', unsafe_allow_html=True)
 st.markdown('<div class="sub-header">Visual search tool for comparable personal injury awards in Ontario</div>', unsafe_allow_html=True)
 
@@ -413,7 +143,10 @@ with st.expander("ℹ️ How to Use This Tool", expanded=False):
     </div>
     """, unsafe_allow_html=True)
 
-# Sidebar for controls
+# =============================================================================
+# SIDEBAR - SEARCH PARAMETERS
+# =============================================================================
+
 with st.sidebar:
     st.header("Search Parameters")
 
@@ -427,16 +160,12 @@ with st.sidebar:
     # CPI Data Upload Section
     with st.expander("📊 Update CPI Data (Optional)", expanded=False):
         st.caption("Upload Bank of Canada CPI data to ensure accurate inflation adjustments")
-
-        # Show current data source
         st.info(f"**Current:** {get_data_source()}")
 
-        # Download current CPI data
         col1, col2 = st.columns(2)
 
         with col1:
             st.markdown("**Download current data:**")
-            # Generate CSV content from current data
             import io
             cpi_data = get_cpi_data()
             csv_buffer = io.StringIO()
@@ -465,14 +194,11 @@ with st.sidebar:
 
         if cpi_file is not None:
             try:
-                # Save uploaded file to data directory
                 BOC_CPI_CSV.parent.mkdir(parents=True, exist_ok=True)
                 with open(BOC_CPI_CSV, 'wb') as f:
                     f.write(cpi_file.getbuffer())
 
-                # Reload data
                 new_data = reload_cpi_data()
-
                 st.success(f"✓ CPI data updated successfully! Now have {len(new_data)} years of data.")
 
             except Exception as e:
@@ -514,16 +240,12 @@ with st.sidebar:
     else:
         st.info("No regions selected - will search all cases")
 
-# Initialize session state
-if 'search_results' not in st.session_state:
-    st.session_state.search_results = None
-if 'analysis_data' not in st.session_state:
-    st.session_state.analysis_data = None
+# =============================================================================
+# MAIN CONTENT - EXPERT REPORT UPLOAD
+# =============================================================================
 
-# Main content area
 st.divider()
 
-# Expert Report Upload Section
 with st.expander("📄 Upload Expert/Medical Report (Optional)", expanded=False):
     st.markdown("""
     Upload a medical or expert report PDF, and the system will automatically extract injuries,
@@ -549,23 +271,16 @@ with st.expander("📄 Upload Expert/Medical Report (Optional)", expanded=False)
     if uploaded_file is not None:
         if st.button("🔍 Analyze Expert Report", type="secondary"):
             with st.spinner("Analyzing expert report..."):
-                # Save uploaded file temporarily
                 with tempfile.NamedTemporaryFile(delete=False, suffix='.pdf') as tmp_file:
                     tmp_file.write(uploaded_file.getvalue())
                     tmp_path = tmp_file.name
 
                 try:
-                    # Analyze the report
-                    analysis = analyze_expert_report(
-                        tmp_path,
-                        use_llm=use_llm
-                    )
-
+                    analysis = analyze_expert_report(tmp_path, use_llm=use_llm)
                     st.session_state.analysis_data = analysis
 
                     st.success("✅ Expert report analyzed successfully!")
 
-                    # Display extracted information
                     st.subheader("Extracted Information")
 
                     if analysis.get('injured_regions'):
@@ -595,9 +310,12 @@ with st.expander("📄 Upload Expert/Medical Report (Optional)", expanded=False)
                     st.error(f"❌ Error analyzing report: {str(e)}")
                     st.info("Try using the manual input fields below instead.")
                 finally:
-                    # Clean up temp file
                     if os.path.exists(tmp_path):
                         os.unlink(tmp_path)
+
+# =============================================================================
+# MAIN CONTENT - SEARCH INPUT
+# =============================================================================
 
 st.divider()
 
@@ -625,14 +343,16 @@ with col2:
     st.subheader("Body Map Reference")
     st.info("Body map visualization coming soon. Use the region selector in the sidebar.")
 
-    # Placeholder for SVG body map
     st.caption("Regions available for selection:")
     st.caption("• Head & Spine (5 regions)")
     st.caption("• Torso (3 regions)")
     st.caption("• Upper Limbs (12 regions)")
     st.caption("• Lower Limbs (12 regions)")
 
-# Search results
+# =============================================================================
+# SEARCH EXECUTION AND RESULTS DISPLAY
+# =============================================================================
+
 if search_button:
     if not injury_text.strip():
         st.warning("⚠️ Please enter an injury description")
@@ -641,11 +361,14 @@ if search_button:
             results = search_cases(
                 injury_text,
                 selected_regions,
+                cases,
+                region_map,
+                model,
                 gender=gender if gender != "Not Specified" else None,
                 age=age
             )
 
-        # Store results in session state for PDF generation
+        # Store results in session state
         st.session_state.search_results = {
             'results': results,
             'injury_text': injury_text,
@@ -665,7 +388,7 @@ if search_button:
             if damage_val:
                 damages_values.append(damage_val)
 
-        # Damage summary
+        # Damage summary metrics
         if damages_values:
             col1, col2, col3 = st.columns(3)
 
@@ -694,122 +417,63 @@ if search_button:
             st.divider()
             st.subheader("📊 Inflation-Adjusted Award Comparison")
 
-            # Prepare chart data
-            chart_data = []
-            for case, emb_sim, combined_score in results[:CHART_MAX_CASES]:
-                damage_val = extract_damages_value(case)
-                year = case.get('year')
-                case_name = case.get('case_name', 'Unknown')
-                citation = case.get('citation', case.get('summary_text', '')[:50])
+            fig = create_inflation_chart(results, DEFAULT_REFERENCE_YEAR)
 
-                if damage_val and year:
-                    adjusted_val = adjust_for_inflation(damage_val, year, DEFAULT_REFERENCE_YEAR)
-                    if adjusted_val:
-                        chart_data.append({
-                            'case_name': case_name,
-                            'year': year,
-                            'original_award': damage_val,
-                            'adjusted_award': adjusted_val,
-                            'citation': citation,
-                            'match_score': combined_score * 100
-                        })
-
-            if chart_data:
-                # Create interactive Plotly chart
-                fig = go.Figure()
-
-                # Add original awards as bars
-                fig.add_trace(go.Bar(
-                    name=f'Original Award',
-                    x=[f"{d['case_name'][:20]}... ({d['year']})" for d in chart_data],
-                    y=[d['original_award'] for d in chart_data],
-                    marker_color='lightblue',
-                    hovertemplate='<b>%{x}</b><br>' +
-                                  'Original Award: $%{y:,.0f}<br>' +
-                                  '<extra></extra>'
-                ))
-
-                # Add inflation-adjusted awards as bars
-                fig.add_trace(go.Bar(
-                    name=f'Inflation-Adjusted ({DEFAULT_REFERENCE_YEAR}$)',
-                    x=[f"{d['case_name'][:20]}... ({d['year']})" for d in chart_data],
-                    y=[d['adjusted_award'] for d in chart_data],
-                    marker_color='darkblue',
-                    customdata=[[d['case_name'], d['citation'], d['match_score'], d['year'],
-                                d['original_award'], d['adjusted_award']] for d in chart_data],
-                    hovertemplate='<b>%{customdata[0]}</b><br>' +
-                                  'Year: %{customdata[3]}<br>' +
-                                  'Original: $%{customdata[4]:,.0f}<br>' +
-                                  f'Adjusted ({DEFAULT_REFERENCE_YEAR}$): $%{{customdata[5]:,.0f}}<br>' +
-                                  'Match Score: %{customdata[2]:.1f}%<br>' +
-                                  'Citation: %{customdata[1]}<br>' +
-                                  '<extra></extra>'
-                ))
-
-                # Update layout
-                fig.update_layout(
-                    barmode='group',
-                    title=f'Damage Awards: Original vs Inflation-Adjusted to {DEFAULT_REFERENCE_YEAR}',
-                    xaxis_title='Case (Year)',
-                    yaxis_title='Award Amount ($)',
-                    yaxis=dict(tickformat='$,.0f'),
-                    hovermode='closest',
-                    height=500,
-                    legend=dict(
-                        orientation="h",
-                        yanchor="bottom",
-                        y=1.02,
-                        xanchor="right",
-                        x=1
-                    ),
-                    xaxis={'categoryorder': 'total descending'}
-                )
-
+            if fig:
                 st.plotly_chart(fig, use_container_width=True)
 
-                # Summary statistics
-                col1, col2, col3 = st.columns(3)
+                # Calculate and display statistics
+                chart_data = []
+                for case, emb_sim, combined_score in results[:CHART_MAX_CASES]:
+                    damage_val = extract_damages_value(case)
+                    year = case.get('year')
+                    if damage_val and year:
+                        from inflation_adjuster import adjust_for_inflation
+                        adjusted_val = adjust_for_inflation(damage_val, year, DEFAULT_REFERENCE_YEAR)
+                        if adjusted_val:
+                            chart_data.append({
+                                'original_award': damage_val,
+                                'adjusted_award': adjusted_val
+                            })
 
-                with col1:
-                    median_original = np.median([d['original_award'] for d in chart_data])
-                    median_adjusted = np.median([d['adjusted_award'] for d in chart_data])
-                    st.metric(
-                        "Median (Original)",
-                        f"${median_original:,.0f}",
-                        help="Median of original award amounts"
+                if chart_data:
+                    stats = calculate_chart_statistics(chart_data)
+
+                    col1, col2, col3 = st.columns(3)
+
+                    with col1:
+                        st.metric(
+                            "Median (Original)",
+                            f"${stats['median_original']:,.0f}",
+                            help="Median of original award amounts"
+                        )
+
+                    with col2:
+                        st.metric(
+                            f"Median ({DEFAULT_REFERENCE_YEAR}$)",
+                            f"${stats['median_adjusted']:,.0f}",
+                            delta=f"+${stats['median_adjusted'] - stats['median_original']:,.0f}",
+                            help=f"Median adjusted to {DEFAULT_REFERENCE_YEAR} dollars"
+                        )
+
+                    with col3:
+                        st.metric(
+                            "Avg. Inflation Impact",
+                            f"+{stats['avg_inflation_impact']:.1f}%",
+                            help="Average percentage increase due to inflation"
+                        )
+
+                    st.caption(
+                        f"💡 **Note:** Awards adjusted to {DEFAULT_REFERENCE_YEAR} dollars using "
+                        "Canadian Consumer Price Index (Statistics Canada). "
+                        "Hover over bars for detailed case information."
                     )
-
-                with col2:
-                    st.metric(
-                        f"Median ({DEFAULT_REFERENCE_YEAR}$)",
-                        f"${median_adjusted:,.0f}",
-                        delta=f"+${median_adjusted - median_original:,.0f}",
-                        help=f"Median adjusted to {DEFAULT_REFERENCE_YEAR} dollars"
-                    )
-
-                with col3:
-                    avg_inflation = np.mean([
-                        ((d['adjusted_award'] - d['original_award']) / d['original_award']) * 100
-                        for d in chart_data
-                    ])
-                    st.metric(
-                        "Avg. Inflation Impact",
-                        f"+{avg_inflation:.1f}%",
-                        help="Average percentage increase due to inflation"
-                    )
-
-                st.caption(
-                    f"💡 **Note:** Awards adjusted to {DEFAULT_REFERENCE_YEAR} dollars using "
-                    "Canadian Consumer Price Index (Statistics Canada). "
-                    "Hover over bars for detailed case information."
-                )
-
             else:
                 st.info("Inflation adjustment requires case year information. Some cases may not have dates.")
 
         st.divider()
 
-        # Display results
+        # Display individual cases
         st.subheader(f"Top {len(results)} Comparable Cases")
 
         for idx, (case, emb_sim, combined_score) in enumerate(results, 1):
@@ -844,11 +508,13 @@ if search_button:
                     if case.get("region_score", 0) > 0:
                         st.metric("Region Match", f"{case['region_score']*100:.0f}%")
 
-# Display results from session state (if available)
+# =============================================================================
+# PDF REPORT GENERATION
+# =============================================================================
+
 if st.session_state.search_results:
     st.divider()
 
-    # PDF Download Section
     st.subheader("📥 Download Report")
 
     col_dl1, col_dl2, col_dl3 = st.columns([1, 1, 2])
@@ -866,25 +532,21 @@ if st.session_state.search_results:
         if st.button("📄 Generate PDF Report", type="secondary"):
             with st.spinner("Generating PDF report..."):
                 try:
-                    # Get search results from session state
                     search_data = st.session_state.search_results
                     results = search_data['results']
 
-                    # Extract damages values
                     damages_values = []
                     for case, emb_sim, combined_score in results:
                         damage_val = extract_damages_value(case)
                         if damage_val:
                             damages_values.append(damage_val)
 
-                    # Create region labels map
                     region_labels = {
                         rid: region_map[rid]["label"]
                         for rid in search_data['selected_regions']
                         if rid in region_map
                     }
 
-                    # Generate PDF
                     timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
                     pdf_filename = f"damages_report_{timestamp}.pdf"
 
@@ -903,13 +565,11 @@ if st.session_state.search_results:
                             max_cases=num_cases
                         )
 
-                        # Read the generated PDF
                         with open(pdf_path, 'rb') as pdf_file:
                             pdf_data = pdf_file.read()
 
                         st.success("✅ PDF report generated successfully!")
 
-                        # Offer download
                         st.download_button(
                             label="💾 Download PDF Report",
                             data=pdf_data,
@@ -925,7 +585,10 @@ if st.session_state.search_results:
     with col_dl3:
         st.caption("Generate a professional PDF report with search parameters, damage analysis, and comparable cases.")
 
-# Footer
+# =============================================================================
+# FOOTER
+# =============================================================================
+
 st.divider()
 st.caption("© 2024 Ontario Damages Compendium Tool | Built for legal professionals | Data source: CCLA Damages Compendium 2024")
 st.caption("⚠️ This tool provides reference information only. Always verify case details and consult primary sources.")
